+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - HasUser trait is now removed due to unintended side effects
 - InternalApiTesting class can no longer be called, use CallsInternalApi trait instead
 - Remove AUTH_PROVIDER env support, use HYVOR_FAKE (internal.fake) instead to mock all auth, billing, and resource
+- FakeProvider $DATABASE is no longer a const, rather a public property
+- FakeProvider renamed to AuthFake
 
 ## 1.1.x - 2024-08-05